+++ src/Plugins.RSP.TAS/Config.h
@@ -6,24 +6,6 @@
 
 #pragma once
 
-#define PLUGIN_VERSION "1.0.0"
-
-#ifdef _M_X64
-#define PLUGIN_ARCH " x64"
-#else
-#define PLUGIN_ARCH " x86"
-#endif
-
-#ifdef _DEBUG
-#define PLUGIN_TARGET " Debug"
-#else
-#define PLUGIN_TARGET " Release"
-#endif
-
-#define PLUGIN_ISAEXT " SSE2"
-
-#define PLUGIN_NAME "TAS RSP " PLUGIN_VERSION PLUGIN_ARCH PLUGIN_ISAEXT PLUGIN_TARGET
-
 struct t_config {
     int32_t version = 2;
     /**
