+++ src/Plugins.RSP.TAS/Resource.h
