--- src/Plugins.RSP.TAS/win.h
@@ -1,18 +1,0 @@
-/*
- * Copyright (c) 2025, Mupen64 maintainers, contributors, and original authors (Hacktarux, ShadowPrince, linker).
- *
- * SPDX-License-Identifier: GPL-2.0-or-later
- */
-
-#pragma once
-
-#include <Core/stdafx.h>
-#include <core_api.h>
-#include <core_plugin.h>
-#include <resource.h>
-
-extern HINSTANCE g_instance;
-extern std::filesystem::path g_app_path;
-extern PlatformService g_platform_service;
-
-INT_PTR CALLBACK ConfigDlgProc(HWND hwnd, UINT Message, WPARAM wParam, LPARAM lParam);
